% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recover_table.R
\name{constraints}
\alias{constraints}
\title{Calculate the matrix constraints}
\usage{
constraints(a, ratio = c("fixed", "sequential"))
}
\arguments{
\item{a}{the matrix}

\item{ratio}{a string specifying the type of ratios used. The default is \code{"fixed"} for ratios with fixed first row and column.
It is also possible to use \code{"sequential"} ratios}
}
\value{
a list with the following elements
\item{cs}{the vector of column sums}
\item{rs}{the vector of row sums without the last element}
\item{p}{the vector of likelihood ratios}
\item{A}{the constraint matrix for sums without the last row}
\item{B}{the constraint matrix for log-likelihood ratios}
}
\description{
Calculate the matrix constraints
}
\examples{
m <- matrix(1:20, nrow = 5)
constraints(m)
}

