% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recover_table.R
\name{cr}
\alias{cr}
\title{Generate the constraints given the vectorized matrix and the constraint matrices}
\usage{
cr(x, p)
}
\arguments{
\item{x}{the vectorized matrix}

\item{p}{the matrix of fixed log ratios#'}
}
\value{
a vector
}
\description{
Generate the constraints given the vectorized matrix and the constraint matrices
}

