% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recover_table.R
\name{cumsum2}
\alias{cumsum2}
\title{Cumulative sum for a matrix}
\usage{
cumsum2(p)
}
\arguments{
\item{p}{a matrix}
}
\value{
a matrix of cumulative sums
}
\description{
If p is initial matrix, the result is the matrix with the following elements:
\deqn{r_{ij} = \sum_{k=1}^i\sum_{l=1}^j p_{ij}}
}

