% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recover_table.R
\name{genA}
\alias{genA}
\title{Generate a constraint matrix for column and row sums}
\usage{
genA(n, m)
}
\arguments{
\item{n}{integer, number of rows}

\item{m}{integer, number of columns}
}
\value{
constraint matrix with n+m rows and n*m columns
}
\description{
Generate a constraint matrix for column and row sums
}

