% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recover_table.R
\name{genB1}
\alias{genB1}
\title{Generate a constraint matrix with fixed first row and column
for bivariate log-likelihood ratio}
\usage{
genB1(n, m)
}
\arguments{
\item{n}{integer, number of rows}

\item{m}{integer, number of columns}
}
\value{
constraint matrix with (n-1)*(m-1) rows and n*m columns
}
\description{
Generate a constraint matrix with fixed first row and column
for bivariate log-likelihood ratio
}

