% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recover_table.R
\name{invp}
\alias{invp}
\title{Recover table from a sequential log ratios and first column and row.}
\usage{
invp(p, col1, row1)
}
\arguments{
\item{p}{a matrix of log-likelihood ratios}

\item{col1}{a vector containing the first column}

\item{row1}{a vector containing the first row}
}
\value{
a matrix
}
\description{
Recover table from a sequential log ratios and first column and row.
}

