% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recover_table.R
\name{jac_cr}
\alias{jac_cr}
\title{Jacobian of the system for recovering the table}
\usage{
jac_cr(x, p, A, exclude = NULL, ...)
}
\arguments{
\item{x}{the vectorized matrix}

\item{p}{the matrix of fixed log ratios}

\item{A}{the constraint matrix for sums}

\item{exclude}{the matrix containing the indices of elements that should be excluded. Default is \code{NULL}, for no exclusions.}

\item{...}{additional parameters which are ignored.}
}
\value{
a matrix
}
\description{
Jacobian of the system for recovering the table
}

