% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recover_table.R
\name{recover_table}
\alias{recover_table}
\title{Recover table given column and row sums and given the log ratios}
\usage{
recover_table(p, col_sums, row_sums, exclude = NULL, ratio = c("fixed",
  "sequential"), ...)
}
\arguments{
\item{p}{the log ratios}

\item{col_sums}{the vector of column sums}

\item{row_sums}{the vector of row sums}

\item{exclude}{the matrix containing the indices of elements that should be excluded. Default is \code{NULL}, for no exclusions.}

\item{ratio}{the type of ratios used. The default is \code{"fixed"} for ratios with fixed first row and column.
It is also possible to use \code{"sequential"} ratios}

\item{...}{additional arguments to nleqlsv}
}
\value{
a list with the following elements
\item{table}{a recovered table}
\item{opt}{a result of \code{nleqslv}}
}
\description{
Recover table given column and row sums and given the log ratios
}
\examples{
set.seed(10)
m <- matrix(runif(9),3)
cr <- constraints(m)
res <- recover_table(cr$p, colSums(m), rowSums(m))
res$table - m
}

