% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recover_table.R
\name{slv}
\alias{slv}
\title{The mismatch between the matrix and its constraints}
\usage{
slv(x, cs, rs, p, A, B)
}
\arguments{
\item{x}{the vectorized matrix}

\item{cs}{the vector of column sums}

\item{rs}{the vector of row sums}

\item{p}{the vector of log-likelihood ratios}

\item{A}{the constraint matrix for sums}

\item{B}{the constraint matrix for log-likelihood ratios}
}
\value{
the vector of differences
}
\description{
The mismatch between the matrix and its constraints
}

