% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recover_table.R
\name{cr}
\alias{cr}
\title{Generate the constraints given the vectorized matrix and the constraint matrices}
\usage{
cr(x, A, B)
}
\arguments{
\item{x}{the vectorized matrix}

\item{A}{the constraint matrix for sums}

\item{B}{the constraint matrix for log-likelihood ratio}
}
\value{
a vector
}
\description{
Generate the constraints given the vectorized matrix and the constraint matrices
}

