% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recover_table.R
\name{jac_cr}
\alias{jac_cr}
\title{Jacobian of the system for recovering the table}
\usage{
jac_cr(x, A, B, ...)
}
\arguments{
\item{x}{the vectorized matrix}

\item{A}{the constraint matrix for sums}

\item{B}{the constraint matrix for log-likelihood ratio}

\item{...}{additional parameters which are ignored.}
}
\value{
a matrix
}
\description{
Jacobian of the system for recovering the table
}

