% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recover_table.R
\name{recover_table}
\alias{recover_table}
\title{Recover table given column and row sums and initial guess}
\usage{
recover_table(initM, col_sums, row_sums, ...)
}
\arguments{
\item{initM}{the initial matrix}

\item{col_sums}{the vector of column sums}

\item{row_sums}{the vector of row sums}

\item{...}{additional arguments to nleqlsv}
}
\value{
a list with the following elements
\item{table}{a recovered table}
\item{opt}{a result of \code{nleqslv}}
}
\description{
Recover table given column and row sums and initial guess
}
\examples{
set.seed(10)
m <- matrix(runif(9),3)
ma <- m + matrix(runif(9),3)/20
res <- recover_table(ma, colSums(m), rowSums(m))
res$table - m
}

