% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recover_table.R
\name{rescale_table}
\alias{rescale_table}
\title{Rescale table given column and row sums and initial table}
\usage{
rescale_table(initM, col_sums, row_sums, ...)
}
\arguments{
\item{initM}{the initial matrix}

\item{col_sums}{the vector of column sums}

\item{row_sums}{the vector of row sums}

\item{...}{additional arguments to nleqlsv}
}
\value{
a list with the following elements
\item{table}{a recovered table}
\item{opt}{a result of \code{nleqslv}}
}
\description{
Rescales table so that the result has the supplied column and row sums. The matrix must
have positive elements in first row and column and nonnegative elements elsewhere.
}
\examples{
set.seed(10)
m <- matrix(runif(9),3)
ma <- m + matrix(runif(9),3)/20
res <- rescale_table(ma, colSums(m), rowSums(m))
res$table - m
ma[3,3] <- 0
res <- rescale_table(ma, colSums(m), rowSums(m))
res$table
}

