% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recover_table.R
\name{slv}
\alias{slv}
\title{The mismatch between the matrix and its constraints}
\usage{
slv(x, cs, rs, p, ...)
}
\arguments{
\item{x}{the vectorized matrix}

\item{cs}{the vector of column sums}

\item{rs}{the vector of row sums}

\item{p}{the matrix of fixed log ratios}

\item{...}{additional parameters which are ignored}
}
\value{
the vector of differences
}
\description{
The mismatch between the matrix and its constraints
}

